ints worden doubles

+++ sonar-polymorphic-views/src/main/java/be/kuleuven/cs/oss/datautils/Position.java
@@ -8,19 +8,19 @@ package be.kuleuven.cs.oss.datautils;
 
 public class Position {
 	
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Position(int x, int y){
+	public Position(double x, double y){
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
